MediaMine/media-mine-database/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
cinemagoer==2022.2.11
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
IMDbPY==2022.7.9
lxml==4.9.1
pytz==2022.2.1
requests==2.28.1
SQLAlchemy==1.4.40
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.11
